Parallelize dropping of `""`
