+++ ecmascript/minifier/src/compress/optimize/strings.rs
@@ -394,56 +394,4 @@ impl Optimizer<'_> {
             _ => {}
         }
     }
-
-    pub(super) fn drop_useless_addition_of_str(&mut self, e: &mut Expr) {
-        match e {
-            Expr::Bin(BinExpr {
-                op: op!(bin, "+"),
-                left,
-                right,
-                ..
-            }) => {
-                let lt = left.get_type();
-                let rt = right.get_type();
-                if let Known(Type::Str) = lt {
-                    if let Known(Type::Str) = rt {
-                        match &**left {
-                            Expr::Lit(Lit::Str(Str {
-                                value: js_word!(""),
-                                ..
-                            })) => {
-                                self.changed = true;
-                                log::debug!(
-                                    "string: Dropping empty string literal (in lhs) because it \
-                                     does not changes type"
-                                );
-
-                                *e = *right.take();
-                                return;
-                            }
-                            _ => {}
-                        }
-
-                        match &**right {
-                            Expr::Lit(Lit::Str(Str {
-                                value: js_word!(""),
-                                ..
-                            })) => {
-                                self.changed = true;
-                                log::debug!(
-                                    "string: Dropping empty string literal (in rhs) because it \
-                                     does not changes type"
-                                );
-
-                                *e = *left.take();
-                                return;
-                            }
-                            _ => {}
-                        }
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
 }
